Backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.1
azure-cognitiveservices-speech==1.45.0
gunicorn==21.2.0